Move tests to a new file
